requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
attrs==24.2.0
bidict==0.23.1
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Markdown==0.3
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
frozenlist==1.5.0
gevent==24.10.3
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2_markdown==0.0.3
kombu==5.4.2
limits==3.13.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
ordered-set==4.1.0
packaging==24.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
redis==5.1.1
requests==2.32.3
rich==13.9.2
rq==2.0.0
rsa==4.9
setuptools==75.3.0
simple-websocket==1.0.0
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.18.3
zope.event==5.0
zope.interface==7.1.1
